Updated to take a better approach to downloading game data
Previously, games.go would only look at game IDs that were
in `game_info` table. This was to make updating faster but
ran the risk of not all stats being downloaded for all players.
Recently ran into this problem; two people were missed on an update
(likely because the game ended just before the loader ran) and then
could not be picked back up as the game was already in `game_info`.

There is now a seperate check to grab all the played games for each
summoner from `summoner_games` and a check to make sure games only
go into `game_info` once.