WIP removed checks that are slowing things down and are unneeded
The plan is to let the database do the majority of the checks by
using an update/insert query instead of the current `INSERT` one.
By removing the search through all of the summoner's games, we can
check against a single ID or datetime and only deal with games that
are after that.

This project should build but it may not work correctly at this time