WIP using channels to provide summoner ids to methods
Decided to switch to channels instead of a `[]int64` so that a
future version could you a fan out approach to download game data
from Riot in a concurrent fashion. In the process, I removed a lot
of code that was no longer needed.

I've also moved the configuration code to a separate file in an
effort to keep `leagueloader.go` clean. It also allows for room
to expand if needed.

This commit will build but will not download game data as I have
temporarily disabled it.