Removed a table and method to reduce time and sycnronization problems
Dropped the `summoner_stats` and related method to reduce having
to synchronize the stats with summoner games. I also removed the
surrogate key on `summoner_games` and replaced it with a composite
key. What this means is that we can simply `INSERT IGNORE...` into
`game_info` and `summoner_games` and not have to worry about
anything. By making the summoner games method use only games after
the game info method is done with them, we can keep the tables
foreign key and data integrity.

The program is currently not running due to trying to get the game
channels to close at the correct time.