Moved the goroutine up one level to surround the summoner loop
In doing this, we're able to have more control over when the channel
closes. It also gets around a previous bug that was causing the wrong
summoner to be assigned a game. The database would end up with more
than 10 games for some summoners, which is impossible on a first run,
and some summoners with less. The summoner ID that was being referenced
was outside the scope of the goroutine and therefore out of its
control.

The application now runs and shuts down properly.